+++ client/components/ManageUsers.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import {Paper} from '@material-ui/core'
+
+const ManageUsers = () => {
+  return (
+    <>
+      <div className="adminContainer">
+        <div className="adminListContainer">
+          <Paper>HELLO THIS IS USERS</Paper>
+        </div>
+      </div>
+    </>
+  )
+}
+
+export default ManageUsers
