+++ client/components/index.js
@@ -9,3 +9,5 @@ export {Login, Signup} from './auth-form'
 export {default as AllProducts} from './AllProducts'
 export {default as SingleProduct} from './SingleProduct'
 export {default as AdminHome} from './AdminHome'
+export {default as ManageProducts} from './ManageProducts'
+export {default as ManageUsers} from './ManageUsers'
