+++ client/store/allProducts.js
@@ -3,6 +3,7 @@ import axios from 'axios'
 const SET_PRODUCTS = 'SET_PRODUCTS'
 const REMOVE_PRODUCT = 'REMOVE_PRODUCT'
 const ADD_PRODUCT = 'ADD_PRODUCT'
+const UPDATE_SINGLE_PRODUCT = `UPDATE_SINGLE_PRODUCT`
 
 export const setProducts = products => {
   return {
@@ -22,6 +23,13 @@ export const addProduct = product => {
   return {
     type: ADD_PRODUCT,
     product
+  }
+}
+
+const updateSingleProduct = product => {
+  return {
+    type: UPDATE_SINGLE_PRODUCT,
+    products
   }
 }
 
@@ -58,6 +66,19 @@ export const postProduct = product => {
   }
 }
 
+export const updateProduct = (productId, productUpdates) => {
+  return async dispatch => {
+    try {
+      await axios.put(`/api/products/${productId}`, productUpdates)
+      const {data} = await axios.get('/api/products')
+      console.log('UPDATED DAT:', data)
+      dispatch(setProducts(data))
+    } catch (err) {
+      console.log(err)
+    }
+  }
+}
+
 export default function productsReducer(state = [], action) {
   switch (action.type) {
     case SET_PRODUCTS:
@@ -68,6 +89,8 @@ export default function productsReducer(state = [], action) {
       )
     case ADD_PRODUCT:
       return [action.product, ...state]
+    // case UPDATE_SINGLE_PRODUCT:
+    //   return [...state].map( product => { product.id === action.product.id ? action.product : product })
     default:
       return state
   }
